put style in components
